[Merge]

+++ branches/avidemux_2.5_branch_gruntster/avidemux/ADM_encoder/adm_encConfig.cpp
@@ -38,7 +38,7 @@ extern struct COMPRES_PARAMS *AllVideoCodec;
 extern int AllVideoCodecCount;
 extern uint8_t DIA_videoCodec(int *codecIndex);
 extern void UI_setVideoCodec(int i);
-extern void getMainWindowHandles(int *handle, int *nativeHandle);
+extern void getMainWindowHandles(long int *handle, long int *nativeHandle);
 
 // Some static stuff
 void setVideoEncoderSettings(COMPRESSION_MODE mode, uint32_t param, uint32_t extraConf, uint8_t *extraData);
