+++ branches/avidemux_2.5_branch_gruntster/avidemux/ADM_plugin/ADM_vidEnc_plugin.h
@@ -87,8 +87,8 @@ typedef struct
 typedef struct
 {
 	int structSize;
-	int parent;
-	int parentNative;
+	long int parent;
+	long int parentNative;
 } vidEncConfigParameters;
 
 int vidEncGetEncoders(int uiType, int **encoderIds);
